Re-organise repo layout to support service definitions
