+++ src/server/api/index.ts
@@ -1,6 +1,6 @@
 import { PubSub } from 'apollo-server'
 import { buildApiServer } from './lib/static_server'
-import { contractServers, availablePorts } from '../storage'
+import { contractServers, availablePorts } from '../infrastructure/storage'
 import { initializeProxy } from './lib/proxy'
 import { Server } from 'http'
 import { ApolloServer } from 'apollo-server-express'
