+++ src/server/api/test/integration.spec.ts
@@ -2,7 +2,7 @@ import fetch from 'cross-fetch'
 import { configureApi } from '..'
 import { ApolloServer, gql } from 'apollo-server'
 import { Server } from 'http'
-import { contractServers } from '../../storage'
+import { contractServers } from '../../infrastructure/storage'
 import { closeAndRemoveServer } from '../lib/contract_servers'
 import { expect } from 'chai'
 import { execute, makePromise, GraphQLRequest } from 'apollo-link'
