+++ src/server/execution_controller/lib/adapter/external.ts
