+++ src/server/execution_controller/lib/adapter/index.ts
